refactor: add MessageParser abstraction
This allows for LegacyParser to be used instead of AdventureParser.
