+++ core/src/main/java/com/rexcantor64/triton/packetinterceptor/handlers/ResourcePackPacketHandler.java
@@ -1,10 +1,9 @@
 package com.rexcantor64.triton.packetinterceptor.handlers;
 
 import com.github.retrooper.packetevents.event.PacketSendEvent;
-import com.github.retrooper.packetevents.wrapper.play.server.WrapperPlayServerDisconnect;
 import com.github.retrooper.packetevents.wrapper.play.server.WrapperPlayServerResourcePackSend;
 import com.rexcantor64.triton.config.MainConfig;
-import com.rexcantor64.triton.language.parser.AdventureParser;
+import com.rexcantor64.triton.language.parser.MessageParser;
 import com.rexcantor64.triton.player.TritonLanguagePlayer;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
@@ -14,10 +13,10 @@ import org.jetbrains.annotations.NotNull;
 @RequiredArgsConstructor
 public class ResourcePackPacketHandler {
 
-    private final @NotNull AdventureParser parser;
+    private final @NotNull MessageParser parser;
     private final @NotNull MainConfig.FeatureSyntax syntax;
 
-    public ResourcePackPacketHandler(@NotNull AdventureParser parser, @NotNull MainConfig config) {
+    public ResourcePackPacketHandler(@NotNull MessageParser parser, @NotNull MainConfig config) {
         this.parser = parser;
         this.syntax = config.getResourcePackPromptSyntax();
     }
